Update README.md
To use this program, run it via a terminal, then input the constants of a quadratic equation in the order they appear in the equation. This program works for equations with the form:

ax^2 + bx + c = 0

where a, b, and c are the constants, and x is the variable. Let the computer do math for you.